requirements.txt
Flask==2.3.2
requests==2.31.0
google-genai==1.15.0
